4.21: blogilistan laajennus, step9 - final

+++ osa4/controllers/blogs.js
@@ -38,11 +38,32 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if(!decodedToken.id) {
+    return response
+      .status(401) /* UNAUTHORIZED */
+      .json({ error: 'token missing or invalid' })
+  }
+  const userId = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(request.params.id)
+
+  if(blog === null) {
+    return response
+      .status(404) /* NOT FOUND */
+      .json({ error: 'blog does not exist' })
+  }
+
+  if(blog.user.toString() !== userId._id.toString()) {
+    return response
+      .status(401) /* UNAUTHORIZED */
+      .json({ error: 'wrong user' })
+  }
 
   await Blog.findByIdAndDelete(request.params.id)
   response
     .status(204) /* NO CONTENT */
     .end()
+
 
 })
